user_management/requirements.txt
Django==3.0.2
djangorestframework==3.10.3
django-webpack-loader==0.6.0
djangorestframework-jwt==1.11
django-cors-headers==3.2.1
coverage==5.0.3
